requirements-debian.txt
clamav-daemon
python3-django
python3-pil
python3-bleach